discord-bot/requirements.txt
discord.py>=2.3,<3.0
SQLAlchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
python-dotenv>=1.0,<2.0
